// frontend/src/lib/generated/api/models/DTORespuestaPaginadaDTOAlumno.ts
export interface DTORespuestaPaginadaDTOAlumno {
    /**
     * List of elements for the current page
     * @type {Array<DTOAlumno>}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    content: Array<DTOAlumno>;
    /**
     * Current page number (0-indexed)
     * @type {number}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    page: number;
    /**
     * Page size
     * @type {number}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    size: number;
    /**
     * Total number of elements
     * @type {number}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    totalElements: number;
    /**
     * Total number of pages
     * @type {number}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    totalPages: number;
    /**
     * Indicates if this is the first page
     * @type {boolean}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    isFirst: boolean;
    /**
     * Indicates if this is the last page
     * @type {boolean}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    isLast: boolean;
    /**
     * Indicates if the page has content
     * @type {boolean}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    hasContent: boolean;
    /**
     * Field by which results are sorted
     * @type {string}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    sortBy: string;
    /**
     * Sort direction
     * @type {string}
     * @memberof DTORespuestaPaginadaDTOAlumno
     */
    sortDirection: string;
}